Move data.rs → data/mod.rs; new data/shared_data.rs
